+++ libs/Utils.py
@@ -1,0 +1,31 @@
+import os
+
+
+def checkPaths(paths):
+    for path in paths:
+        if os.path.exists(path):
+            log("found path: %s" % path)
+            return path
+    return ""
+
+
+def findWord(view):
+    for region in view.sel():
+        if region.begin() == region.end():
+            word = view.word(region)
+        else:
+            word = region
+        if not word.empty():
+            return view.substr(word)
+        else:
+            return ""
+
+
+def jump_to_label_declaration(view, label_id):
+    view.run_command("insert", {"characters": label_id})
+    view.hide_popup()
+
+
+def log(string):
+    print("SublimeKodi: " + string)
+
